+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 Changelog
 =========
 
+v0.4.0
+------
+
+* #3: Implement 'NextTicket' and 'PrevTicket'
+* Add :create {project} {summary} command
+* #31 - use age template function
+* #8 - Allow breaking out of the UI to run external commands (eg vim)
+* Add various quick search commands
+* Add 'previousPages history', to make 'q' less painful
+* Swtich to use ScrollableList base interface widget
+
 v0.3.2
 ------